+++ lib/rails/generators/component/component_generator.rb
@@ -6,15 +6,18 @@ module Rails
       source_root File.expand_path("templates", __dir__)
 
       argument :attributes, type: :array, default: [], banner: "attribute"
-      hook_for :test_framework
       check_class_collision suffix: "Component"
+
+      class_option :require_content, type: :boolean, default: false
 
       def create_component_file
         template "component.rb", File.join("app/components", class_path, "#{file_name}_component.rb")
       end
 
-      def create_template_file
-        template "component.html.erb", File.join("app/components", class_path, "#{file_name}_component.html.erb")
+      hook_for :test_framework
+
+      hook_for :template_engine do |instance, template_engine|
+        instance.invoke template_engine, [instance.name], require_content: instance.send(:requires_content?)
       end
 
       private
@@ -24,6 +27,7 @@ module Rails
       end
 
       def requires_content?
+        return if behavior == :revoke
         return @requires_content if @asked
 
         @asked = true